Moved waf and wafscript.
Now they all reside in /shared/DSPFilters/Builds/GNULinux.

Note that ./waf needs to be executed from that directory even when one
wants to build DSPFiltersDemo. This solution is not really nice, but
avoids cluttering the whole repository with wscripts.

The canonical solution would be placing a wscript into all the
directories under /shared and a wscript in the project root than use
recursion. But as long as waf is only used for the GNU/Linux build, this
is definitely an overkill.

Also added a quick hack to make the JUCE amalgam compile under g++ 4.7,
it seems <unistd.h> does not get included automatically any more.
The wscript basically passes -include unistd.h to the compiler when
needed.